requirements.txt
torch>=1.8.1
numpy>=1.19.5
biopython>=1.78
joblib>=1.0.1
matplotlib>=3.3.4
scikit-learn>=0.24.1